use cargo_bin!()
Closes: https://github.com/sayanarijit/xplr/issues/768
