+++ .github/workflows/nix-cache.yml
@@ -9,11 +9,11 @@ jobs:
         os: [ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.os }}
     steps:
-      - uses: actions/checkout@v3
-      - uses: nixbuild/nix-quick-install-action@v19
+      - uses: actions/checkout@v4
+      - uses: nixbuild/nix-quick-install-action@v21
         with:
           nix_conf: experimental-features = nix-command flakes
-      - uses: cachix/cachix-action@v11
+      - uses: cachix/cachix-action@v12
         with:
           name: xplr
           authtoken: "${{ secrets.CACHIX_AUTH_TOKEN }}"
